tests: Add test for stream memory leaking
The H3Connection object stores each stream in a dictionary and never
clears it, which leads to a memory leak. Before fixing it, let's add
a test case.

Though the test itself is a bit hacky because it accesses internal
of the H3Connection object. I hope it's okay, because I don't see
any other way of asserting the number of stored streams, except maybe
by introducing a public API for this.